+++ tests/gaussian/test_jobs.py
@@ -37,8 +37,9 @@ class TestGaussianJob(TestCase):
         os.chdir(CWD)
         shutil.rmtree(SCR_DIR)
         files_to_remove = glob.glob(f"{TEST_DIR}/*.out")
-        for file_path in files_to_remove:
-            os.remove(file_path)
+        if files_to_remove and glob.glob(f"{TEST_DIR}/*.out.gz"):
+            for file_path in files_to_remove:
+                os.remove(file_path)
 
     def test_normal(self):
         g = GaussianJob(
@@ -51,8 +52,9 @@ class TestGaussianJob(TestCase):
         )
         g.setup()
         assert os.path.exists(f"{SCR_DIR}/test.com.orig")
-        with gzip.open(f"{TEST_DIR}/mol_opt.out.gz", "rb") as f_in, open(f"{TEST_DIR}/mol_opt.out", "wb") as f_out:
-            shutil.copyfileobj(f_in, f_out)
+        if not os.path.exists(f"{TEST_DIR}/mol_opt.out") and os.path.exists(f"{TEST_DIR}/mol_opt.out.gz"):
+            with gzip.open(f"{TEST_DIR}/mol_opt.out.gz", "rb") as f_in, open(f"{TEST_DIR}/mol_opt.out", "wb") as f_out:
+                shutil.copyfileobj(f_in, f_out)
         shutil.copy(f"{TEST_DIR}/mol_opt.out", f"{SCR_DIR}/test.out")
         g.postprocess()
         assert os.path.exists(f"{SCR_DIR}/test.com{self.suffix}")
@@ -72,8 +74,9 @@ class TestGaussianJob(TestCase):
         assert len(jobs) == 1, "One job should be generated under normal conditions."
         jobs[0].setup()
         assert os.path.exists(f"{SCR_DIR}/test.com.orig")
-        with gzip.open(f"{TEST_DIR}/mol_opt.out.gz", "rb") as f_in, open(f"{TEST_DIR}/mol_opt.out", "wb") as f_out:
-            shutil.copyfileobj(f_in, f_out)
+        if not os.path.exists(f"{TEST_DIR}/mol_opt.out") and os.path.exists(f"{TEST_DIR}/mol_opt.out.gz"):
+            with gzip.open(f"{TEST_DIR}/mol_opt.out.gz", "rb") as f_in, open(f"{TEST_DIR}/mol_opt.out", "wb") as f_out:
+                shutil.copyfileobj(f_in, f_out)
         shutil.copy(f"{TEST_DIR}/mol_opt.out", f"{SCR_DIR}/test.out")
         jobs[0].postprocess()
         assert os.path.exists(f"{SCR_DIR}/test.com.guess1")
